Add files via upload
JSON Files in Python

More JSON in Python
For the example in this lesson, JSON data was sourced from an API. That isn't always the case, though! Sometimes you're given a text file with human readable JSON within it. For this situation, the json library is indispensable. It can parse JSON from strings or files and it can parse JSON into a Python dictionary or list. It can also convert Python dictionaries or lists into JSON strings. The tutorial on the linked documentation page is handy. This Reading and Writing JSON to a File in Python article from Stack Abuse is also great, which outlines json.dump, json.dumps, json.load, and json.loads (four key json library methods) well.

pandas also has JSON functions (the read_json function and the to_json DataFrame method), but the hierarchical advantage of JSON is wasted in pandas' tabular DataFrame so the uses are limited.